Moved CdiVaadinServiceDelegate as a nested class of CdiVaadinServletService
